+++ index.php
@@ -6,7 +6,6 @@ if(isset($_SESSION['admin']))
 }
 
 include 'mc-core/loader.php';
-
 //Function operatives
  if (isset($_GET['page']))
  {
@@ -20,13 +19,13 @@ elseif (isset($_POST['page']))
  {
 	 $d = NULL;
  }
-
 // Error Checking
-$core->check_errors(true);
-
+if (defined('DEBUG', 'TRUE'))
+{
+   $core->check_errors(true);
+}
  // Include the theme's Header
 include 'mc-content/themes/'.$theme.'/header.php';
-
 /*
  * Prevention Session Injection
  */
@@ -34,18 +33,14 @@ if (isset($_REQUEST['_SESSION']))
 {
 	die("Get lost Muppet!");
 }
-
 // Set the Desired Time Zone
 date_default_timezone_set($timezone);
-
-
 $query = "SELECT * FROM mc_plugins WHERE plugin_status = 1";
 $result = $db->query($query);
 $finished = false;
 while($row = $db->fetch_assoc($result))
 {
     $plugin = $row['plugin_slug'];
-	
     if ($d === $plugin)
 	{
         include 'mc-content/plugins/'.$plugin.'/'.$plugin.'.php';
@@ -63,6 +58,12 @@ while($row = $db->fetch_assoc($result))
 	if ($d === 'error')
 	{
 		include 'mc-includes/core_plugins/error/error.php';
+		$finished = true;
+		break;
+	}
+	if ($d === 'privacy-policy')
+	{
+		include 'mc-includes/core_plugins/privacy-policy/privacy-policy.php';
 		$finished = true;
 		break;
 	}
@@ -103,12 +104,6 @@ if (!$finished)
     include 'mc-content/themes/'.$theme.'/front_page.php';
 }
 
-
-// Do SSL and Sentry check
-include 'mc-includes/footer-check.php';
-
-
 // Include the themes Footer
 include 'mc-content/themes/'.$theme.'/footer.php';
-
 ?>